rename wrapper to WorkflowWrapper
