+++ src/pages/workflows/workflow/WorkflowWdl.js
@@ -7,7 +7,7 @@ import * as breadcrumbs from 'src/components/breadcrumbs';
 import WDLViewer from 'src/components/WDLViewer';
 import { forwardRefWithName } from 'src/libs/react-utils';
 import { snapshotStore } from 'src/libs/state';
-import { wrapWorkflows } from 'src/pages/workflows/workflow/Wrapper';
+import { wrapWorkflows } from 'src/pages/workflows/workflow/WorkflowWrapper';
 
 export const BaseWorkflowWdl = () => {
   const { name, snapshotId, payload } = useStore(snapshotStore);
